chore(properties): move camunda properties to correct provider
* We mixed up the ErrorEvent properties from bpmn and camunda context
* This moves the camunda related to the right spot
