+++ lib/provider/bpmn/parts/EventProps.js
@@ -61,13 +61,8 @@ module.exports = function(group, element, bpmnFactory, elementRegistry, translat
       var errorEventDefinition = eventDefinitionHelper.getErrorEventDefinition(element);
 
       if (errorEventDefinition) {
-        var isCatchingErrorEvent = is(element, 'bpmn:StartEvent') || is (element, 'bpmn:BoundaryEvent');
 
-        var showErrorCodeVariable = isCatchingErrorEvent,
-            showErrorMessageVariable = isCatchingErrorEvent;
-
-        error(group, element, bpmnFactory, errorEventDefinition, showErrorCodeVariable, showErrorMessageVariable,
-          translate);
+        error(group, element, bpmnFactory, errorEventDefinition, translate);
       }
     }
   });
