+++ lib/provider/camunda/CamundaPropertiesProvider.js
@@ -26,6 +26,7 @@ var serviceTaskDelegateProps = require('./parts/ServiceTaskDelegateProps'),
     multiInstanceProps = require('./parts/MultiInstanceLoopProps'),
     conditionalProps = require('./parts/ConditionalProps'),
     scriptProps = require('./parts/ScriptTaskProps'),
+    errorProps = require('./parts/ErrorEventProps'),
     formProps = require('./parts/FormProps'),
     startEventInitiator = require('./parts/StartEventInitiator'),
     variableMapping = require('./parts/VariableMappingProps'),
@@ -187,6 +188,7 @@ function createGeneralTabGroups(
   linkProps(detailsGroup, element, translate);
   callActivityProps(detailsGroup, element, bpmnFactory, translate);
   eventProps(detailsGroup, element, bpmnFactory, elementRegistry, translate);
+  errorProps(detailsGroup, element, bpmnFactory, translate);
   conditionalProps(detailsGroup, element, bpmnFactory, translate);
   startEventInitiator(detailsGroup, element, translate); // this must be the last element of the details group!
