+++ test/spec/provider/bpmn/ErrorEventDefinitionSpec.js
@@ -61,14 +61,6 @@ function getErrorCodeField(container) {
   return getInputField(container, 'error-element-code', 'errorCode');
 }
 
-function getErrorCodeVariableField(container) {
-  return getInputField(container, 'errorCodeVariable', 'errorCodeVariable');
-}
-
-function getErrorMessageVariableField(container) {
-  return getInputField(container, 'errorMessageVariable', 'errorMessageVariable');
-}
-
 function getClearButton(container, selector) {
   return domQuery('div[data-entry=' + selector + '] button[data-action=clear]', container);
 }
@@ -162,24 +154,6 @@ describe('error-event-properties', function() {
       var field = getErrorCodeField(container);
 
       expect(field.value).to.equal(errorEventDefinition.errorRef.errorCode);
-
-    });
-
-
-    it('should fetch the errorCodeVariable of an error event', function() {
-
-      var field = getErrorCodeVariableField(container);
-
-      expect(field.value).to.equal(errorEventDefinition.get('camunda:errorCodeVariable'));
-
-    });
-
-
-    it('should fetch the errorMessageVariable of an error event', function() {
-
-      var field = getErrorMessageVariableField(container);
-
-      expect(field.value).to.equal(errorEventDefinition.get('camunda:errorMessageVariable'));
 
     });
 
@@ -331,156 +305,6 @@ describe('error-event-properties', function() {
 
           // then
           expect(errorEventDefinition.errorRef.errorCode).to.be.undefined;
-        }));
-
-      });
-
-    });
-
-
-    describe('should set the errorCodeVariable for an error event', function() {
-
-      var errorEventDefinition, field;
-
-      beforeEach(inject(function(elementRegistry, selection, propertiesPanel) {
-
-        // given
-        var item = elementRegistry.get('ErrorStartEvent_Empty');
-        selection.select(item);
-
-        var bo = item.businessObject;
-        errorEventDefinition = eventDefinitionHelper.getErrorEventDefinition(bo);
-
-        field = getErrorCodeVariableField(container);
-
-        // when
-        TestHelper.triggerValue(field, 'FOO', 'change');
-
-      }));
-
-      describe('in the DOM', function() {
-
-        it('should execute', function() {
-          expect(field.value).to.equal('FOO');
-        });
-
-        it('should undo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-
-          // then
-          expect(field.value).to.equal('');
-        }));
-
-
-        it('should redo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-          commandStack.redo();
-
-          // then
-          expect(field.value).to.equal('FOO');
-        }));
-
-      });
-
-      describe('on the business object', function() {
-
-        it('should execute', function() {
-          expect(errorEventDefinition.get('camunda:errorCodeVariable')).to.equal('FOO');
-        });
-
-        it('should undo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-
-          // then
-          expect(errorEventDefinition.get('camunda:errorCodeVariable')).to.be.undefined;
-        }));
-
-
-        it('should redo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-          commandStack.redo();
-
-          // then
-          expect(errorEventDefinition.get('camunda:errorCodeVariable')).to.equal('FOO');
-        }));
-
-      });
-
-    });
-
-
-    describe('should set the errorMessageVariable for an error event', function() {
-
-      var errorEventDefinition, field;
-
-      beforeEach(inject(function(elementRegistry, selection, propertiesPanel) {
-
-        // given
-        var item = elementRegistry.get('ErrorStartEvent_Empty');
-        selection.select(item);
-
-        var bo = item.businessObject;
-        errorEventDefinition = eventDefinitionHelper.getErrorEventDefinition(bo);
-
-        field = getErrorMessageVariableField(container);
-
-        // when
-        TestHelper.triggerValue(field, 'FOO', 'change');
-
-      }));
-
-      describe('in the DOM', function() {
-
-        it('should execute', function() {
-          expect(field.value).to.equal('FOO');
-        });
-
-        it('should undo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-
-          // then
-          expect(field.value).to.equal('');
-        }));
-
-
-        it('should redo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-          commandStack.redo();
-
-          // then
-          expect(field.value).to.equal('FOO');
-        }));
-
-      });
-
-      describe('on the business object', function() {
-
-        it('should execute', function() {
-          expect(errorEventDefinition.get('camunda:errorMessageVariable')).to.equal('FOO');
-        });
-
-        it('should undo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-
-          // then
-          expect(errorEventDefinition.get('camunda:errorMessageVariable')).to.be.undefined;
-        }));
-
-
-        it('should redo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-          commandStack.redo();
-
-          // then
-          expect(errorEventDefinition.get('camunda:errorMessageVariable')).to.equal('FOO');
         }));
 
       });
@@ -780,188 +604,6 @@ describe('error-event-properties', function() {
         }));
 
       });
-
-    });
-
-    describe('#errorCodeVariable', function() {
-
-      var field, button;
-
-      beforeEach(function() {
-
-        field = getErrorCodeVariableField(container);
-        button = getClearButton(container, 'errorCodeVariable');
-
-        TestHelper.triggerEvent(button, 'click');
-
-      });
-
-      describe('in the DOM', function() {
-
-        it('should execute', function() {
-          expect(field.value).is.empty;
-        });
-
-        it('should undo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-
-          // then
-          expect(field.value).to.equal('myErrorCodeVariable');
-        }));
-
-
-        it('should redo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-          commandStack.redo();
-
-          // then
-          expect(field.value).is.empty;
-        }));
-
-      });
-
-      describe('on the business object', function() {
-
-        it('should execute', function() {
-          expect(errorEventDefinition.get('camunda:errorCodeVariable')).to.be.undefined;
-        });
-
-        it('should undo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-
-          // then
-          expect(errorEventDefinition.get('camunda:errorCodeVariable')).to.equal('myErrorCodeVariable');
-        }));
-
-
-        it('should redo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-          commandStack.redo();
-
-          // then
-          expect(errorEventDefinition.get('camunda:errorCodeVariable')).to.be.undefined;
-        }));
-
-      });
-
-    });
-
-
-    describe('#errorMessageVariable', function() {
-
-      var field, button;
-
-      beforeEach(function() {
-
-        field = getErrorMessageVariableField(container);
-        button = getClearButton(container, 'errorMessageVariable');
-
-        TestHelper.triggerEvent(button, 'click');
-
-      });
-
-      describe('in the DOM', function() {
-
-        it('should execute', function() {
-          expect(field.value).is.empty;
-        });
-
-        it('should undo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-
-          // then
-          expect(field.value).to.equal('myErrorMessageVariable');
-        }));
-
-
-        it('should redo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-          commandStack.redo();
-
-          // then
-          expect(field.value).is.empty;
-        }));
-
-      });
-
-      describe('on the business object', function() {
-
-        it('should execute', function() {
-          expect(errorEventDefinition.get('camunda:errorMessageVariable')).to.be.undefined;
-        });
-
-        it('should undo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-
-          // then
-          expect(errorEventDefinition.get('camunda:errorMessageVariable')).to.equal('myErrorMessageVariable');
-        }));
-
-
-        it('should redo', inject(function(commandStack) {
-          // when
-          commandStack.undo();
-          commandStack.redo();
-
-          // then
-          expect(errorEventDefinition.get('camunda:errorMessageVariable')).to.be.undefined;
-        }));
-
-      });
-
-    });
-
-  });
-
-
-  describe('control visibility', function() {
-
-    function expectVisible(elementId, visible, getter) {
-
-      return inject(function(propertiesPanel, selection, elementRegistry) {
-
-        // given
-        var element = elementRegistry.get(elementId);
-
-        // assume
-        expect(element).to.exist;
-
-        // when
-        selection.select(element);
-        var field = getter(propertiesPanel._container);
-
-        // then
-        if (visible) {
-          expect(field).to.exist;
-        } else {
-          expect(field).not.to.exist;
-        }
-      });
-    }
-
-
-    describe('should show', function() {
-
-      it('StartEvent_ErrorCodeVariable', expectVisible('ErrorStartEvent_Empty', true, getErrorCodeVariableField));
-      it('StartEvent_ErrorMessageVariable', expectVisible('ErrorStartEvent_Empty', true, getErrorMessageVariableField));
-
-      it('BoundaryEvent_ErrorCodeVariable', expectVisible('ErrorBoundaryEvent_All', true, getErrorCodeVariableField));
-      it('BoundaryEvent_ErrorMessageVariable', expectVisible('ErrorBoundaryEvent_All', true, getErrorMessageVariableField));
-
-    });
-
-
-    describe('should hide', function() {
-
-      it('EndEvent_ErrorCodeVariable', expectVisible('ErrorEndEvent_ErrorCode', false, getErrorCodeVariableField));
-      it('EndEvent_ErrorMessageVariable', expectVisible('ErrorEndEvent_ErrorCode', false, getErrorMessageVariableField));
 
     });
