+++ test/spec/provider/camunda/ErrorEventDefinition.bpmn
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<bpmn:definitions xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:camunda="http://camunda.org/schema/1.0/bpmn" id="Definitions_1" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="3.3.0-dev">
+  <bpmn:process id="Process_1" isExecutable="false">
+    <bpmn:endEvent id="ErrorEndEvent_ErrorMessage" name="ErrorEndEvent_ErrorMessage">
+      <bpmn:errorEventDefinition errorRef="Error_1" />
+    </bpmn:endEvent>
+    <bpmn:startEvent id="ErrorStartEvent_all" name="ErrorStartEvent_all">
+      <bpmn:errorEventDefinition errorRef="Error_1" camunda:errorCodeVariable="myErrorCodeVariable" camunda:errorMessageVariable="myErrorMessageVariable" />
+    </bpmn:startEvent>
+  </bpmn:process>
+  <bpmn:error id="Error_1" name="b" errorCode="code" camunda:errorMessage="message" />
+  <bpmndi:BPMNDiagram id="BPMNDiagram_1">
+    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="Process_1">
+      <bpmndi:BPMNShape id="EndEvent_11kkznz_di" bpmnElement="ErrorEndEvent_ErrorMessage">
+        <dc:Bounds x="222" y="79" width="36" height="36" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="197" y="115" width="86" height="27" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="StartEvent_0lwts7h_di" bpmnElement="ErrorStartEvent_all">
+        <dc:Bounds x="503" y="79" width="36" height="36" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="477" y="122" width="89" height="27" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNShape>
+    </bpmndi:BPMNPlane>
+  </bpmndi:BPMNDiagram>
+</bpmn:definitions>
